Began breaking out the UmbracoContext into a base class and creating specific implementations where needed
